Update README.md
Working with Type Casting

Run this program and notice the results in the console window. There is a logic error in this program centering around integer division.

This program is similar to Lab 4.2. This time, do NOT change the data type of the two named constants to get the correct result. Instead, use a typecast to solve the problem. The correct result should be 0.292162. Work with this program until you get the correct result.